backend/requirements.txt
fastapi==0.115.13
uvicorn==0.34.3
torch==2.7.1
python-multipart==0.0.20
python-dotenv==1.0.1

# Whisper from GitHub
openai-whisper @ git+https://github.com/openai/whisper.git
